feat: Discord AutoMod Extension (#315)
fix: extra and invalid events in the action log

+++ src/events/discord/GuildMemberRoleRemove.ts
@@ -1,25 +1,22 @@
-import { Collection, GuildMember, Role } from 'discord.js'
+import { GuildMember, Role } from 'discord.js'
 import Lacuna from '../../internals/Lacuna'
 import Automation from '../../modules/Automation'
-import Logs from '../../modules/Logs'
 
-const handler = async (self: Lacuna, member: GuildMember, roles: Collection<string, Role>) => {
+const handler = async (self: Lacuna, member: GuildMember, roles: Role[]) => {
     const server = await self.db.servers.fetch({ _id: member.guild.id })
+    if (!server || server.blocked) return false
 
     await Automation.handleEvent('ROLE_MEMBER_REMOVE', self, server, member)
 
     const temprole = self.temproles.find(i => i.user_id === member.id)
-
     if (temprole && roles.some(i => i.id === temprole.role_id)) {
         await temprole.delete()
     }
-
-    await Logs.RoleMemberRemove(self, server, member, roles)
 
     return true
 }
 
 export default {
-    name: 'roleMemberRemove',
+    name: 'guildMemberRoleRemove',
     handler
 }
